Refactor Mail class: move to the top level to be used with both sandbox and sending APIs
