+++ lib/mailtrap.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative 'mailtrap/action_mailer' if defined? ActionMailer
+require_relative 'mailtrap/mail'
 require_relative 'mailtrap/sending'
 require_relative 'mailtrap/version'
