+++ spec/mailtrap/mail/base_spec.rb
@@ -3,7 +3,7 @@
 require_relative './shared'
 
 # rubocop:disable RSpec/MultipleMemoizedHelpers
-RSpec.describe Mailtrap::Sending::Mail do
+RSpec.describe Mailtrap::Mail::Base do
   subject(:mail) do
     described_class.new(
       from: from,
@@ -124,16 +124,16 @@ RSpec.describe Mailtrap::Sending::Mail do
     let(:category) { 'another_category' }
     let(:expected_json) do
       '{' \
-        '"to":[{"email":"to@example.com"},{"email":"to2@example.com","name":"To Two"}],' \
         '"from":{"email":"test@example.com","name":"Mailtrap User"},' \
+        '"to":[{"email":"to@example.com"},{"email":"to2@example.com","name":"To Two"}],' \
         '"cc":[{"email":"cc@example.com"}],' \
         '"bcc":[{"email":"bcc@example.com"}],' \
+        '"subject":"This is subject",' \
+        '"text":"This is text",'\
+        '"html":"<div>Test HTML</div>",' \
         '"attachments":[{"content":"aGVsbG8gd29ybGQ=","filename":"attachment.txt"}],' \
         '"headers":{"Category-Header":"some_category"},' \
         '"custom_variables":{},' \
-        '"subject":"This is subject",' \
-        '"html":"<div>Test HTML</div>",' \
-        '"text":"This is text",'\
         '"category":"another_category"' \
         '}'
     end
