+++ spec/mailtrap/mail/shared.rb
